infra/hunyuan_requirements.txt
fastapi
uvicorn[standard]
pillow
torch==2.2.2+cu121 torchvision==0.17.2+cu121 --extra-index-url https://download.pytorch.org/whl/cu121
transformers
accelerate
google-cloud-storage